Added a new schedule for creating the picks page
Also, refactored the jobs from 'bots' to 'jobs' package
