--- app/api/templates/game_update_flow.yml
@@ -1,40 +1,0 @@
-id: $flow_id
-namespace: $namespace
-tasks:
-  - id: update_game_loop
-    type: io.kestra.plugin.scripts.python.Script
-    timeout: "PT6H"
-    script: |
-      import requests
-      import time
-      requests.packages.urllib3.disable_warnings(requests.packages.urllib3.exceptions.InsecureRequestWarning)
-
-      game_status = 'STATUS_SCHEDULED'
-      while game_status != 'STATUS_FINAL':
-        response = requests.get(
-          url="$uri",
-          headers={
-            "Authorization": "Bearer {{ secret('TGFP_API_TOKEN') }}"
-          },
-        )
-        game = response.json()
-        print(game)
-        game_status = game['game_status']
-        if game_status != 'STATUS_FINAL':
-          time.sleep(300)
-        else:
-          print("JHS: Game is final, waiting two minutes, then updating the team record")
-          time.sleep(120)
-          response = requests.post(
-            url="$update_uri",
-            headers={
-              "Authorization": "Bearer {{ secret('TGFP_API_TOKEN') }}"
-            },
-          )
-
-triggers:
-  - id: schedule
-    type: io.kestra.plugin.core.trigger.Schedule
-    timezone: US/Pacific
-    cron: "$cron"
-
