--- app/api/templates/nag_player_flow.yml
@@ -1,16 +1,0 @@
-id: $flow_id
-namespace: $namespace
-tasks:
-  - id: nag_players
-    type: io.kestra.plugin.core.http.Request
-    uri: $uri
-    method: POST
-    headers:
-      authorization: "Bearer {{ secret('TGFP_API_TOKEN') }}"
-    contentType: application/json
-triggers:
-  - id: schedule
-    type: io.kestra.plugin.core.trigger.Schedule
-    timezone: US/Pacific
-    cron: "$cron"
-
